Feat: Added a REST API to fetch batteries in provided range with the corresponding statistics on power (Wattage)

+++ src/main/java/com/proshore/assessment/dto/BatteryDto.java
@@ -42,5 +42,5 @@ public class BatteryDto implements Serializable {
      */
     @NotNull
     @Positive
-    private Double capacity;
+    private Integer capacity;
 }
